# board-service/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateProjectJoinRequestRequest:
    properties:
      project_id:
        type: string
    required:
    - project_id
    type: object
  dto.CreateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      workspace_id:
        type: string
    required:
    - name
    - workspace_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
      request_id:
        type: string
    type: object
...
  dto.PaginatedProjectsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      projects:
        items:
          $ref: '#/definitions/dto.ProjectResponse'
        type: array
      total:
        type: integer
    type: object
  dto.ProjectJoinRequestResponse:
    properties:
      project_id:
        type: string
      request_id:
        type: string
      requestedAt:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
      userEmail:
        type: string
      userName:
        type: string
    type: object
  dto.ProjectMemberResponse:
    properties:
      joinedAt:
        type: string
      member_id:
        type: string
      project_id:
        type: string
      roleName:
        type: string
      user_id:
        type: string
      userEmail:
        type: string
      userName:
        type: string
    type: object
  dto.ProjectResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      name:
        type: string
      owner_id:
        type: string
      ownerEmail:
        type: string
      ownerName:
        type: string
      project_id:
        type: string
      updatedAt:
        type: string
      workspace_id:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
    type: object
...
  dto.UpdateProjectJoinRequestRequest:
    properties:
      status:
        enum:
        - APPROVED
        - REJECTED
        type: string
    required:
    - status
    type: object
  dto.UpdateProjectMemberRoleRequest:
    properties:
      roleName:
        enum:
        - OWNER
        - ADMIN
        - MEMBER
        type: string
    required:
    - roleName
    type: object
  dto.UpdateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@wealist.com
    name: API Support
  description: Board management API for weAlist project management platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Board Service API
  version: "1.0"
paths:
...
  /api/projects:
    get:
      consumes:
      - application/json
      description: Get all projects in a workspace (workspace member only)
      parameters:
      - description: Workspace ID
        in: query
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProjectResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project in a workspace (workspace member only)
      parameters:
      - description: Project details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /api/projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a project (OWNER only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get project details (project member only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update project details (OWNER only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Updated project details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /api/projects/{project_id}/join-requests:
    get:
      consumes:
      - application/json
      description: Get join requests for a project (OWNER/ADMIN only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Filter by status (PENDING/APPROVED/REJECTED)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProjectJoinRequestResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get join requests
      tags:
      - projects
  /api/projects/{project_id}/members:
    get:
      consumes:
      - application/json
      description: Get all members of a project (member only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProjectMemberResponse'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project members
      tags:
      - projects
  /api/projects/{project_id}/members/{member_id}:
    delete:
      consumes:
      - application/json
      description: Remove a member from project (OWNER/ADMIN only, cannot remove OWNER
        or self)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove member
      tags:
      - projects
  /api/projects/{project_id}/members/{member_id}/role:
    put:
      consumes:
      - application/json
      description: Update a member's role in project (OWNER only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProjectMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectMemberResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - projects
...
  /api/projects/join-requests:
    post:
      consumes:
      - application/json
      description: Request to join a project (workspace member only)
      parameters:
      - description: Join request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectJoinRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectJoinRequestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create join request
      tags:
      - projects
  /api/projects/join-requests/{join_request_id}:
    put:
      consumes:
      - application/json
      description: Approve or reject a join request (OWNER/ADMIN only)
      parameters:
      - description: Join Request ID
        in: path
        name: join_request_id
        required: true
        type: string
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProjectJoinRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProjectJoinRequestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update join request
      tags:
      - projects
  /api/projects/search:
    get:
      consumes:
      - application/json
      description: Search projects in a workspace by name or description
      parameters:
      - description: Workspace ID
        in: query
        name: workspace_id
        required: true
        type: string
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedProjectsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search projects
      tags:
      - projects
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"